# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/files": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "files"
                ],
                "summary": "Get all files on the server",
                "operationId": "getFiles",
                "responses": {
                    "200": {
                        "description": "Files available",
                        "schema": {
                            "$ref": "#/definitions/api.ResponseFiles"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ResponseFiles"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "files"
                ],
                "summary": "Add file",
                "operationId": "addFile",
                "parameters": [
                    {
                        "type": "file",
                        "description": "File to be uploaded",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "File that was added",
                        "schema": {
                            "$ref": "#/definitions/api.ResponseFile"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/api.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ResponseError"
                        }
                    }
                }
            }
        },
        "/files/{fileID}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "files"
                ],
                "summary": "Get file info",
                "operationId": "getFile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of file",
                        "name": "fileID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "File info",
                        "schema": {
                            "$ref": "#/definitions/api.ResponseFile"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/api.ResponseError"
                        }
                    },
                    "404": {
                        "description": "File not found",
                        "schema": {
                            "$ref": "#/definitions/api.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ResponseError"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "files"
                ],
                "summary": "Update file",
                "operationId": "updateFile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of file",
                        "name": "fileID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "File to be uploaded",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "File that was updated",
                        "schema": {
                            "$ref": "#/definitions/api.ResponseFile"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/api.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ResponseError"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "files"
                ],
                "summary": "Delete file",
                "operationId": "deleteFile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of file",
                        "name": "fileID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Succeeds whether the file exists or not",
                        "schema": {
                            "$ref": "#/definitions/api.ResponseEmpty"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ResponseError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.ResponseEmpty": {
            "description": "Empty successful response",
            "type": "object",
            "required": [
                "data"
            ],
            "properties": {
                "data": {
                    "type": "object"
                }
            }
        },
        "api.ResponseError": {
            "description": "An error",
            "type": "object",
            "required": [
                "error"
            ],
            "properties": {
                "error": {
                    "$ref": "#/definitions/api.ResponseErrorData"
                }
            }
        },
        "api.ResponseErrorData": {
            "type": "object",
            "required": [
                "code",
                "message"
            ],
            "properties": {
                "code": {
                    "description": "HTTP response status code",
                    "type": "integer"
                },
                "message": {
                    "description": "Description of the error",
                    "type": "string"
                }
            }
        },
        "api.ResponseFile": {
            "description": "A single file",
            "type": "object",
            "required": [
                "data"
            ],
            "properties": {
                "data": {
                    "$ref": "#/definitions/api.ResponseFileData"
                }
            }
        },
        "api.ResponseFileData": {
            "type": "object",
            "required": [
                "fileID",
                "lastModified"
            ],
            "properties": {
                "fileID": {
                    "description": "ID of file",
                    "type": "string"
                },
                "lastModified": {
                    "description": "Last modified timestamp of file",
                    "type": "string"
                },
                "url": {
                    "description": "URL of file",
                    "type": "string"
                }
            }
        },
        "api.ResponseFiles": {
            "description": "Multiple files (NOTE: Partial responses are possible. In this case `data` and `error` will be returned.)",
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.ResponseFileData"
                    }
                },
                "error": {
                    "$ref": "#/definitions/api.ResponseErrorData"
                }
            }
        }
    }
}